+++ app/scripts/components/wordpicture/results-word-picture.ts
@@ -9,6 +9,7 @@ import "@/components/util/json_button"
 import "@/components/util/korp-error"
 import "@/components/wordpicture/word-picture"
 import { StoreService } from "@/services/store"
+import { CsvType, downloadCsvFile } from "@/csv"
 
 type ResultsWordPictureController = IController & {
     isActive: boolean
@@ -19,6 +20,7 @@ type ResultsWordPictureController = IController & {
 type ResultsWordPictureScope = IScope & {
     activated: boolean
     data?: WordPicture
+    downloadOption: CsvType | ""
     error?: string
     limit: string // Number as string to work with <select ng-model>
     limitOptions: number[]
@@ -75,6 +77,12 @@ angular.module("korpApp").component("resultsWordPicture", {
         </help-box>
 
         <div class="mt-4 flex gap-4 justify-end">
+            <select ng-show="!$ctrl.loading && data && !warning && !error" ng-model="downloadOption">
+                <option value="">{{ "download" | loc:$root.lang }}</option>
+                <option value="comma">{{ "csv_comma" | loc:$root.lang }}</option>
+                <option value="tab">{{ "csv_tab" | loc:$root.lang }}</option>
+            </select>
+
             <json-button
                 ng-if="!$ctrl.loading && data && !warning && !error"
                 endpoint="relations"
@@ -94,6 +102,7 @@ angular.module("korpApp").component("resultsWordPicture", {
         function ($scope: ResultsWordPictureScope, $timeout: ITimeoutService, store: StoreService) {
             const $ctrl = this as ResultsWordPictureController
             $scope.activated = false
+            $scope.downloadOption = ""
             $scope.limit = String(LIMITS[0])
             $scope.limitOptions = [...LIMITS]
             $scope.proxy = new RelationsProxy()
@@ -185,6 +194,19 @@ angular.module("korpApp").component("resultsWordPicture", {
                     })
                     .finally(() => $timeout(() => $ctrl.setProgress(false, 0)))
             }
+
+            // Create and download CSV file when the download selector is used
+            $scope.$watch("downloadOption", () => {
+                // Skip if empty (at init or if the label option is selected)
+                if (!$scope.downloadOption) return
+
+                if (!$scope.data) throw new Error("Word picture data missing")
+                const rows = $scope.data.generateCsv()
+                downloadCsvFile("word-picture", rows, $scope.downloadOption)
+
+                // Reset to the label option
+                $scope.downloadOption = ""
+            })
         },
     ],
 })