+++ app/scripts/word-picture.ts
@@ -117,5 +117,19 @@ export class WordPicture {
     }
 
     /** Get a string for the params that identify a word picture column */
-    getColumnId = (pos: string, rel: string, reverse: boolean) => `${pos}${reverse ? "+" : "-"}${rel}`
+    getColumnId = (pos: string, rel: string, reverse: boolean) => `${pos}${reverse ? "+" : "-"}${rel}`;
+
+    /** Create listing of full data, suitable for CSV export. */
+    *generateCsv(): Generator<string[]> {
+        const fields: (keyof MatchedRelation)[] = ["head", "headpos", "dep", "deppos", "rel", "depextra", "freq", "mi"]
+        // Header row
+        yield fields
+
+        // Data rows
+        for (const key of Object.keys(this.items)) {
+            for (const relation of this.items[key]) {
+                yield fields.map((field) => String(relation[field] || ""))
+            }
+        }
+    }
 }